Refactor to use one Mongo Connection
