+++ .env.sample
@@ -4,7 +4,6 @@ WEBHOOK_TOKEN=
 CLIENT_ID=
 NEWS_API_KEY=
 MONGODB_URI=
-QUIZ_MONGODB_URI=
 AI_KEY=
 B_ID=
 SUGGESTION_WEBHOOK_ID=
